requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
bcrypt==5.0.0
    # via passlib
cachetools==6.2.6
    # via pyiceberg
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   pyiceberg
    #   uvicorn
colorama==0.4.6
    # via click
cryptography==46.0.4
    # via
    #   pyjwt
    #   python-jose
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via pydantic
fastapi==0.128.0
    # via -r requirements.in
fsspec==2026.1.0
    # via pyiceberg
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via pyiceberg
multidict==6.7.1
    # via yarl
packaging==26.0
    # via deprecation
passlib[bcrypt]==1.7.4
    # via -r requirements.in
postgrest==2.27.3
    # via supabase
propcache==0.4.1
    # via yarl
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pycparser==3.0
    # via cffi
pydantic[email]==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   postgrest
    #   pydantic-settings
    #   pyiceberg
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via rich
pyiceberg==0.10.0
    # via storage3
pyjwt[crypto]==2.11.0
    # via supabase-auth
pyparsing==3.3.2
    # via pyiceberg
pyroaring==1.0.3
    # via pyiceberg
python-dateutil==2.9.0.post0
    # via strictyaml
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.22
    # via -r requirements.in
realtime==2.27.3
    # via supabase
requests==2.32.5
    # via pyiceberg
rich==14.3.2
    # via pyiceberg
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sortedcontainers==2.4.0
    # via pyiceberg
starlette==0.50.0
    # via fastapi
storage3==2.27.3
    # via supabase
strenum==0.4.15
    # via supabase-functions
strictyaml==1.7.3
    # via pyiceberg
supabase==2.27.3
    # via -r requirements.in
supabase-auth==2.27.3
    # via supabase
supabase-functions==2.27.3
    # via supabase
tenacity==9.1.2
    # via pyiceberg
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   realtime
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via -r requirements.in
websockets==15.0.1
    # via realtime
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-functions
